chore: Version core auth CRDs
